+++ pickupShuttle/src/main/java/pickup_shuttle/pickup/_core/errors/exception/Exception500.java
@@ -2,7 +2,6 @@ package pickup_shuttle.pickup._core.errors.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
-import pickup_shuttle.pickup._core.utils.ApiResult;
 import pickup_shuttle.pickup._core.utils.ApiUtils;
 
 // 서버 에러
@@ -12,7 +11,7 @@ public class Exception500 extends RuntimeException {
         super(message);
     }
 
-    public ApiResult<?> body(){
+    public ApiUtils.ApiResult<?> body(){
         return ApiUtils.error(getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
     }
