--- pickupShuttle/src/main/java/pickup_shuttle/pickup/_core/utils/ApiError.java
@@ -1,4 +1,0 @@
-package pickup_shuttle.pickup._core.utils;
-
-public record ApiError(String message, int status) {
-}
