--- pickupShuttle/src/main/java/pickup_shuttle/pickup/_core/utils/ApiResult.java
@@ -1,4 +1,0 @@
-package pickup_shuttle.pickup._core.utils;
-
-public record ApiResult<T>(boolean success, T response, ApiError error) {
-}
