+++ pickupShuttle/src/main/java/pickup_shuttle/pickup/domain/user/User.java
@@ -1,4 +1,4 @@
-package pickup_shuttle.pickup.domain;
+package pickup_shuttle.pickup.domain.user;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
