+++ pickupShuttle/src/main/java/pickup_shuttle/pickup/domain/user/UserService.java
@@ -1,4 +1,4 @@
-package pickup_shuttle.pickup.domain;
+package pickup_shuttle.pickup.domain.user;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
