+++ pickupShuttle/src/test/java/pickup_shuttle/pickup/domain/user/UserControllerTest.java
@@ -9,8 +9,6 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
-import pickup_shuttle.pickup.domain.UserRequest;
-import pickup_shuttle.pickup.domain.board.BoardService;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
